+++ test/casedb/integration/refdata_access/test_casedb_edge_cases_refdata_access.py
@@ -12,10 +12,11 @@ from gen_epix.fastapp import CrudOperation
 from gen_epix.seqdb.domain import enum as seqdb_enum
 from test.casedb.casedb_test_client import CasedbTestClient as Env
 
-from test.casedb.integration.conftest import (
+from test.casedb.integration.define_edge_cases import (
     EDGE_CASES,
     EdgeCaseSpec,
 )
+
 from test.casedb.integration.refdata_access.base_empty import (
     DEV_REPOSITORY_CONFIG,
     SKIP_ENDPOINTS,
@@ -158,6 +159,9 @@ class TestCaseDBEdgeCasesRefDataAccess:
             f"\n  Unexpected access: {sorted(unexpected) if unexpected else '∅'}"
         )
 
+    # Note: This one uses the setup_case_type_data that is also used for the case type and case type set access tests
+    # we have to use an additional fixture with specific setup for the case type col sets
+    # because the logic is more complex and we want to keep the expected access sets in the EdgeCaseSpec for clarity and maintainability
     @pytest.mark.parametrize(
         "spec",
         EDGE_CASES,
